python_browser_agent/requirements.txt
# Use recent compatible versions
browser-use>=0.1.40
langchain-core>=0.1.52
langchain-ollama>=0.1.8
langchain_community>=0.3.20
pydantic>=2.0
flask>=2.0.0
flask-socketio>=5.0.0
python-dotenv>=0.19.0
requests>=2.26.0
python-ollama>=0.1.0 
